Auth conf implemented - CRUD Admins implemented
