+++ next-auth.d.ts
@@ -1,18 +1,10 @@
 import 'next-auth';
 
-interface Role {
-  id: number;
-  name_role: string;
-  createdAt: string;
-  updatedAt: string;
-}
-
 declare module 'next-auth' {
   interface User {
-    id: number;
     document: string;
-    name: string;
-    role: Role;
+    fullName: string;
+    role: string;
     token?: string;
   }
