+++ src/app/dashboard/users/page.tsx
@@ -1,14 +1,14 @@
-import CreateUserDialog from '@/components/users-table/create-user-dialog';
-import UserTable from '@/components/users-table/user-table';
-
-export default async function UsersPage() {
+export default function UsersPage() {
   return (
-    <div>
-      <div className='flex items-center justify-between'>
-        <h1 className='py-6 text-xl font-bold'>Manejo de usuarios</h1>
-        <CreateUserDialog />
+    <div className='flex flex-1 flex-col gap-4'>
+      <div className='grid auto-rows-min gap-4 md:grid-cols-3'>
+        <div className='aspect-video rounded-xl bg-muted/50'>
+          <p>hola users</p>
+        </div>
+        <div className='aspect-video rounded-xl bg-muted/50' />
+        <div className='aspect-video rounded-xl bg-muted/50' />
       </div>
-      <UserTable />
+      <div className='min-h-[100vh] flex-1 rounded-xl bg-muted/50 md:min-h-min' />
     </div>
   );
 }
