+++ src/components/admins-table/admin-table.tsx
@@ -3,32 +3,33 @@
 import { DataTable } from '../ui/data-table';
 import { columns } from './columns';
 import { useState } from 'react';
-import { getUsers } from '@/services/userService';
+import { getAdmins } from '@/services/adminService';
 import { getSession } from 'next-auth/react';
 import { useQuery } from '@tanstack/react-query';
 import TableSkeleton from '../table-skeleton';
 
-export default function UserTable() {
+export default function AdminTable() {
   const [pageIndex, setPageIndex] = useState(0);
   const [pageSize, setPageSize] = useState(10);
 
   const { data, isLoading, isError } = useQuery({
-    queryKey: ['users', pageIndex, pageSize],
+    queryKey: ['admins', pageIndex, pageSize],
     queryFn: async () => {
       const session = await getSession();
       const token = session?.user.access_token;
 
       if (!token) return { data: [], totalPages: 0 };
-      const res = await getUsers(token, {
+      const res = await getAdmins(token, {
         page: pageIndex + 1,
         limit: pageSize,
       });
-      return res.UserPaginated;
+      return res;
     },
-    staleTime: 5000,
+    staleTime: 60000,
   });
 
-  if (isError) return <div>Error loading users</div>;
+  if (isError) return <div>Error loading admins</div>;
+  console.log(data);
 
   return (
     <div>
@@ -37,12 +38,12 @@ export default function UserTable() {
       ) : (
         <DataTable
           columns={columns}
-          data={data?.data || []}
+          data={data || []}
           pageIndex={pageIndex}
           setPageIndex={setPageIndex}
           pageSize={pageSize}
           setPageSize={setPageSize}
-          totalPages={data?.totalPages || 0}
+          totalPages={Math.ceil(data?.length / pageSize)}
         />
       )}
     </div>
