--- src/components/dialog-delete.tsx
@@ -1,27 +1,0 @@
-import {
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-} from '@/components/ui/alert-dialog';
-
-export default function DialogDelete() {
-  return (
-    <AlertDialogContent>
-      <AlertDialogHeader>
-        <AlertDialogTitle>¿Estás seguro?</AlertDialogTitle>
-        <AlertDialogDescription>
-          Esta acción no se puede deshacer. Esto eliminará permanentemente el
-          elemento seleccionado.
-        </AlertDialogDescription>
-      </AlertDialogHeader>
-      <AlertDialogFooter>
-        <AlertDialogCancel>Cancelar</AlertDialogCancel>
-        <AlertDialogAction>Eliminar</AlertDialogAction>
-      </AlertDialogFooter>
-    </AlertDialogContent>
-  );
-}
