+++ src/lib/sidebar-items.ts
@@ -3,23 +3,27 @@ import { Bot, HeartHandshake, User2, Users } from 'lucide-react';
 export const sidebarItems = {
   navMain: [
     {
-      title: 'Usuarios',
+      title: 'Gestores y Admins',
       icon: User2,
       isActive: true,
       items: [
         {
-          title: 'Manejar Usuarios',
-          url: '/dashboard/users',
+          title: 'Manejar Gestores',
+          url: '/dashboard/managers',
+        },
+        {
+          title: 'Manejar Administradores',
+          url: '/dashboard/admins',
         },
       ],
     },
     {
-      title: 'Gestores',
+      title: 'Usuarios',
       icon: Bot,
       items: [
         {
-          title: 'Manejar Gestores',
-          url: '/dashboard/managers',
+          title: 'Manejar Usuarios',
+          url: '/dashboard/users',
         },
       ],
     },
