+++ src/middleware.ts
@@ -1,5 +1,5 @@
 export { default } from 'next-auth/middleware';
 
 export const config = {
-  matcher: [],
+  matcher: ['/dashboard', '/dashboard/:path*'],
 };
