+++ src/types/create-admin.ts
@@ -1,0 +1,6 @@
+export interface CreateAdmin {
+  document: string;
+  fullName: string;
+  password: string;
+  email: string;
+}
