--- src/types/create-user.ts
@@ -1,10 +1,0 @@
-export interface CreateUser {
-  document: string;
-  name: string;
-  lastName: string;
-  password: string;
-  status: boolean;
-  role: string;
-  direction: string;
-  email: string;
-}
