+++ src/types/update-admin.ts
@@ -1,0 +1,5 @@
+export interface UpdateAdmin {
+  document: string;
+  fullName: string;
+  email: string;
+}
